--- lib/hydrate.js
@@ -1,45 +1,0 @@
-var some = Array.prototype.some;
-
-function hydrate(frag, callbacks, scope) {
-  var paths = Object.keys(callbacks);
-  var id = +paths.shift();
-  var cur = 0;
-
-  traverse(frag);
-
-  function check(node) {
-    cur++;
-    if(id === cur) {
-      var callback = callbacks[id];
-      callback(node, scope);
-      id = +paths.shift();
-    }
-    return !id;
-  }
-
-  function traverse(node){
-    var exit;
-    some.call(node.attributes || [], function(){
-      exit = check(node);
-      if(exit) {
-        return true;
-      }
-    });
-    if(exit) return false;
-
-    some.call(node.childNodes, function(child){
-      exit = check(child);
-      if(exit) {
-        return true;
-      }
-
-      exit = !traverse(child);
-      if(exit) {
-        return true;
-      }
-    });
-    return !exit;
-  }
-}
-
-export default hydrate;
