+++ lib/index.js
@@ -1,26 +1,8 @@
-const { Readable, Transform, Writable } = require('stream');
+const { Readable } = require('stream');
 const isPromise = require('is-promise');
 const isStream = require('is-stream');
-
-const toString = val => {
-  if(Array.isArray(val)) {
-    return val.join('');
-  }
-  return val.toString();
-};
-
-function pipeInto(fromStream, readable) {
-  let outStream = fromStream.pipe(new Writable({
-    write(data, enc, next) {
-      readable.push(data);
-      next(null, data);
-    }
-  }));
-
-  return new Promise(resolve => {
-    outStream.on('finish', () => resolve());
-  });
-}
+const pipeInto = require('./util/pipe-into');
+const toString = require('./util/to-string');
 
 async function readAll(stream, strings, values) {
   let i = 0;
@@ -60,34 +42,4 @@ exports.html = function(strings, ...values){
   return readable;
 };
 
-function mapStream(stream, cb) {
-  let outStream = new Transform();
-  outStream._transform = function(val, enc, next){
-    next(null, val);
-  };
-
-  let writable = new Writable({
-    objectMode: true,
-    write(data, enc, next) {
-      let val = cb(data);
-      if(isStream(val)) {
-        val.pipe(outStream, { end: false });
-        val.on('error', next);
-        val.on('end', () => next(null, data));
-      } else {
-        // TODO
-        throw new Error('Using map with non-stream entries is not currently supported.');
-      }
-    }
-  });
-
-  writable.on('finish', () => {
-    outStream.end();
-  });
-
-  stream.pipe(writable, { end: true });
-
-  return outStream;
-}
-
-exports.map = mapStream;
+exports.map = require('./map').map;
