--- lib/inspect.js
@@ -1,84 +1,0 @@
-import live from './live.js';
-import parse from './expression.js';
-
-var forEach = Array.prototype.forEach;
-
-function inspect(node, ref, paths) {
-  var ignoredAttrs = {};
-
-  switch(node.nodeType) {
-    // Element
-    case 1:
-      var templateAttr;
-      if(node.nodeName === 'TEMPLATE' && (templateAttr = specialTemplateAttr(node))) {
-        var result = parse(node.getAttribute(templateAttr));
-        if(result.hasBinding) {
-          result.throwIfMultiple();
-          ignoredAttrs[templateAttr] = true;
-          paths[ref.id] = function(node, model){
-            if(templateAttr === 'each') {
-              live.each(node, model, result, node);
-            } else {
-              setupBinding(model, result, live[templateAttr](node, model));
-            }
-          };
-        }
-      }
-      break;
-    // TextNode
-    case 3:
-      var result = parse(node.nodeValue);
-      if(result.hasBinding) {
-        paths[ref.id] = function(node, model){
-          live.text(node, result, model);
-        };
-      }
-      break;
-  }
-
-  forEach.call(node.attributes || [], function(attrNode){
-    // TODO see if this is important
-    ref.id++;
-
-    if(ignoredAttrs[attrNode.name])
-      return;
-
-    var name = attrNode.name;
-    var property = isPropAttr(name);
-    var result = parse(attrNode.value);
-    if(result.hasBinding) {
-      paths[ref.id] = function(node, model){
-        if(property) {
-          throw new Error('Properties not yet implemented');
-        }
-        live.attr(node, result, name, model);
-      };
-    } else if(property) {
-      console.log('still do this');
-    }
-  });
-
-  var childNodes = node.childNodes;
-  forEach.call(childNodes, function(node){
-    ref.id++;
-    inspect(node, ref, paths);
-  });
-
-  return paths;
-}
-
-var specialTemplateAttrs = ['if', 'each'];
-function specialTemplateAttr(template){
-  var attrName;
-  for(var i = 0, len = specialTemplateAttrs.length; i < len; i++) {
-    attrName = specialTemplateAttrs[i];
-    if(template.getAttribute(attrName))
-      return attrName;
-  }
-}
-
-function isPropAttr(name) {
-  return name && name[0] === ':';
-}
-
-export default inspect;
