--- lib/live.js
@@ -1,152 +1,0 @@
-var live = {
-  attr: function(node, result, attrName, model){
-    let val = result.value(model);
-    node.setAttribute(attrName, val);
-  },
-  text: function(node, result, model){
-    var val = result.value(model);
-    node.nodeValue = val;
-  },
-  prop: function(node, prop){
-    return function(val){
-      node[prop] = val;
-    };
-  },
-  each: function(node, parentScope, parseResult){
-    var hydrate = Bram.template(node);
-    var prop = parseResult.props()[0];
-    var scopeResult = parentScope.read(prop);
-    var placeholder = document.createTextNode('');
-    node.parentNode.replaceChild(placeholder, node);
-
-    var observe = function(list){
-      var itemMap = new Map();
-      var indexMap = new Map();
-
-      var render = function(item, i){
-        var scope = parentScope.add(item).add({ item: item, index: i});
-        var frag = hydrate(scope);
-
-        var info = {
-          item: item,
-          nodes: slice.call(frag.childNodes),
-          scope: scope,
-          index: i
-        };
-        itemMap.set(item, info);
-        indexMap.set(i, info);
-
-        var siblingInfo = indexMap.get(i + 1);
-        var parent = placeholder.parentNode;
-        if(siblingInfo) {
-          var firstChild = siblingInfo.nodes[0];
-          parent.insertBefore(frag, firstChild);
-        } else {
-          parent.appendChild(frag);
-        }
-      };
-
-      var remove = function(index){
-        var info = indexMap.get(index);
-        if(info) {
-          info.nodes.forEach(function(node){
-            node.parentNode.removeChild(node);
-          });
-          itemMap.delete(info.item);
-          indexMap.delete(index);
-        }
-      };
-
-      list.forEach(render);
-
-      var onarraychange = function(ev, value){
-        if(ev.type === 'delete') {
-          remove(ev.index);
-          return;
-        }
-
-        var info = itemMap.get(value);
-        if(info) {
-          var oldIndex = info.index;
-          var hasChanged = oldIndex !== ev.index;
-          if(hasChanged) {
-            info.scope.model.index = info.index = ev.index;
-
-            var existingItem = indexMap.get(ev.index);
-            if(existingItem) {
-              indexMap.set(oldIndex, existingItem);
-            } else {
-              indexMap.delete(oldIndex);
-            }
-            indexMap.set(ev.index, info);
-
-            var ref = indexMap.get(ev.index + 1);
-            if(ref) {
-              ref = ref.nodes[0];
-            }
-
-            var nodeIdx = info.nodes.length - 1;
-            while(nodeIdx >= 0) {
-              placeholder.parentNode.insertBefore(info.nodes[nodeIdx], ref);
-              nodeIdx--;
-            }
-          }
-        } else {
-          remove(ev.index);
-          render(value, ev.index);
-        }
-      };
-
-      Bram.addEventListener(list, Bram.arrayChange, onarraychange);
-
-      return function(){
-        for(var i = 0, len = list.length; i < len; i++) {
-          remove(i);
-        }
-        Bram.removeEventListener(list, Bram.arrayChange, onarraychange);
-        itemMap = null;
-        indexMap = null;
-      };
-    };
-
-    var teardown = observe(scopeResult.value);
-
-    Bram.addEventListener(scopeResult.model, prop, function(ev, newValue){
-      teardown();
-      teardown = observe(newValue);
-    });
-  },
-  if: function(node, parentScope){
-    var hydrate = Bram.template(node);
-    var rendered = false;
-    var child = {};
-    var placeholder = document.createTextNode('');
-    node.parentNode.replaceChild(placeholder, node);
-    return function(val){
-      if(!rendered) {
-        if(val) {
-          var scope = parentScope.add(val);
-          var frag = hydrate(scope);
-          child.children = slice.call(frag.childNodes);
-          child.scope = scope;
-          placeholder.parentNode.insertBefore(frag, placeholder.nextSibling);
-          rendered = true;
-        }
-      } else {
-        var parent = placeholder.parentNode;
-        var sibling = placeholder.nextSibling;
-        if(val) {
-          child.children.forEach(function(node){
-            parent.insertBefore(node, sibling);
-          });
-        } else {
-          child.children.forEach(function(node){
-            parent.removeChild(node);
-          });
-        }
-      }
-    };
-  }
-};
-
-export default live;
