+++ lib/map.js
@@ -1,28 +1,37 @@
-let MyMap;
+const isStream = require('is-stream');
+const { Transform, Writable } = require('stream');
 
-if(typeof Map === 'function') {
-  MyMap = Map;
-} else {
-  myMap = function(){
-    this._obj = {};
+function mapStream(stream, cb) {
+  let outStream = new Transform();
+  outStream._transform = function(val, enc, next){
+    next(null, val);
   };
 
-  MyMap.prototype.set = function(key, value){
-    this._obj[key] = value;
-  };
+  let writable = new Writable({
+    objectMode: true,
+    write(data, enc, next) {
+      let val = cb(data);
+      if(isStream(val)) {
+        val.pipe(outStream, { end: false });
+        val.on('error', next);
+        val.on('end', () => next(null, data));
+      } else {
+        // TODO
+        throw new Error('Using map with non-stream entries is not currently supported.');
+      }
+    }
+  });
 
-  MyMap.prototype.get = function(key){
-    return this._obj[key];
-  };
+  writable.on('finish', () => {
+    outStream.end();
+  });
 
-  MyMap.prototype.forEach = function(cb){
-    var map = this;
-    var obj = this._obj;
-    Object.keys(obj).forEach(function(key){
-      var value = obj[key];
-      cb(value, key, map);
-    });
-  };
+  stream.pipe(writable, { end: true });
+
+  return outStream;
 }
 
-export default MyMap;
+exports.mapStream = mapStream;
+
+// TODO make this work with regular arrays and stuff
+exports.map = mapStream;
