+++ lib/util/to-string.js
@@ -1,0 +1,8 @@
+function toString(val) {
+  if(Array.isArray(val)) {
+    return val.join('');
+  }
+  return val.toString();
+}
+
+module.exports = toString;
