+++ package.json
@@ -14,8 +14,7 @@
   "license": "BSD-2-Clause",
   "dependencies": {
     "is-promise": "^2.1.0",
-    "is-stream": "^1.1.0",
-    "is-stream-ended": "^0.1.3"
+    "is-stream": "^1.1.0"
   },
   "devDependencies": {
     "mocha": "^3.1.0",
