--- test/ex.html
@@ -1,3 +1,0 @@
-<template>
-  <span class="msg">Hello <strong>{{name}}</strong>!</span>
-</template>
