+++ test/test.js
@@ -2,20 +2,7 @@ const assert = require('assert');
 const {html, map} = require('../lib/index');
 const fs = require('fs');
 const Readable = require('stream').Readable;
-const through = require('through2');
-
-async function readAll(stream) {
-  let values = [];
-  return new Promise((resolve, reject) => {
-    stream.on('error', err => reject(err));
-    stream.pipe(through(function(val, enc, next){
-      values.push(val.toString().trim());
-      next();
-    }, function(){
-      resolve(values);
-    }));
-  });
-}
+const { readAll } = require('./helpers');
 
 describe('TextNodes', function(){
   it('basics works', async function(){
